CommonUtil: more moving
* Minesweeper stuff to Minesweeper
* JsonDiff stuff to JsonDiff
* Delete deprecated constructors
