+++ projects/batfish-common-protocol/src/main/java/org/batfish/common/util/JsonDiff.java
@@ -2,14 +2,17 @@ package org.batfish.common.util;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
+import java.util.Collection;
 import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.SortedMap;
 import java.util.SortedSet;
 import java.util.TreeMap;
 import java.util.TreeSet;
+import java.util.function.Supplier;
 import org.batfish.common.BatfishException;
 import org.codehaus.jettison.json.JSONArray;
 import org.codehaus.jettison.json.JSONException;
@@ -121,9 +124,9 @@ public class JsonDiff {
         String key = (String) i.next();
         rhsKeys.add(key);
       }
-      SortedSet<String> commonKeys = CommonUtil.intersection(lhsKeys, rhsKeys, TreeSet::new);
-      SortedSet<String> lhsOnlyKeys = CommonUtil.difference(lhsKeys, rhsKeys, TreeSet::new);
-      SortedSet<String> rhsOnlyKeys = CommonUtil.difference(rhsKeys, lhsKeys, TreeSet::new);
+      SortedSet<String> commonKeys = intersection(lhsKeys, rhsKeys, TreeSet::new);
+      SortedSet<String> lhsOnlyKeys = difference(lhsKeys, rhsKeys, TreeSet::new);
+      SortedSet<String> rhsOnlyKeys = difference(rhsKeys, lhsKeys, TreeSet::new);
       for (String lhsOnlyKey : lhsOnlyKeys) {
         String removedKeyName = REMOVED_ITEM_CODE + lhsOnlyKey;
         _data.put(removedKeyName, getValue(lhs.get(lhsOnlyKey)));
@@ -196,6 +199,22 @@ public class JsonDiff {
   @JsonCreator
   private JsonDiff(SortedMap<String, Object> data) {
     _data = data;
+  }
+
+  private static <S extends Set<T>, T> S difference(
+      Set<T> minuendSet, Set<T> subtrahendSet, Supplier<S> setConstructor) {
+    S differenceSet = setConstructor.get();
+    differenceSet.addAll(minuendSet);
+    differenceSet.removeAll(subtrahendSet);
+    return differenceSet;
+  }
+
+  private static <S extends Set<T>, T> S intersection(
+      Set<T> set1, Collection<T> set2, Supplier<S> setConstructor) {
+    S intersectionSet = setConstructor.get();
+    intersectionSet.addAll(set1);
+    intersectionSet.retainAll(set2);
+    return intersectionSet;
   }
 
   @JsonValue
