+++ projects/minesweeper/src/main/java/org/batfish/minesweeper/smt/PropertyChecker.java
@@ -41,8 +41,6 @@ import org.batfish.datamodel.Prefix;
 import org.batfish.datamodel.StaticRoute;
 import org.batfish.datamodel.answers.AnswerElement;
 import org.batfish.datamodel.questions.smt.EnvironmentType;
-import org.batfish.datamodel.questions.smt.HeaderLocationQuestion;
-import org.batfish.datamodel.questions.smt.HeaderQuestion;
 import org.batfish.minesweeper.CommunityVar;
 import org.batfish.minesweeper.Graph;
 import org.batfish.minesweeper.GraphEdge;
@@ -56,6 +54,8 @@ import org.batfish.minesweeper.answers.SmtManyAnswerElement;
 import org.batfish.minesweeper.answers.SmtOneAnswerElement;
 import org.batfish.minesweeper.answers.SmtReachabilityAnswerElement;
 import org.batfish.minesweeper.collections.Table2;
+import org.batfish.minesweeper.question.HeaderLocationQuestion;
+import org.batfish.minesweeper.question.HeaderQuestion;
 import org.batfish.minesweeper.utils.PathRegexes;
 import org.batfish.minesweeper.utils.PatternUtils;
 import org.batfish.minesweeper.utils.TriFunction;
