+++ projects/minesweeper/src/main/java/org/batfish/minesweeper/utils/PathRegexes.java
@@ -3,7 +3,7 @@ package org.batfish.minesweeper.utils;
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
 import org.batfish.common.BatfishException;
-import org.batfish.datamodel.questions.smt.HeaderLocationQuestion;
+import org.batfish.minesweeper.question.HeaderLocationQuestion;
 
 public class PathRegexes {
